pyproject.toml
[project]
name = "workshop-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "simple-salesforce>=1.12.8",
    "snowflake-connector-python>=3.17.2",
]
